# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReCredita API",
    "description": "Recuperação (Plano A) e aquisição (Plano B) de veículos.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "Erro": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string"
          }
        },
        "required": [
          "erro"
        ],
        "additionalProperties": false
      },
...
      "Criado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
...
      "ValorFipe": {
        "type": "object",
        "properties": {
          "valor": {
            "type": "number"
          },
          "codigoFipe": {
            "type": "string"
          },
          "mesReferencia": {
            "type": "string"
          },
          "marca": {
            "type": "string"
          },
          "modelo": {
            "type": "string"
          },
          "anoModelo": {
            "type": "number"
          }
        },
        "required": [
          "valor",
          "codigoFipe",
          "mesReferencia",
          "marca",
          "modelo",
          "anoModelo"
        ],
        "additionalProperties": false
      },
      "VerificacaoVeicular": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "fornecedor": {
            "nullable": true,
            "type": "string"
          },
          "baseLegal": {
            "nullable": true,
            "type": "string"
          },
          "justificativa": {
            "nullable": true,
            "type": "string"
          },
          "situacao": {
            "nullable": true,
            "type": "string"
          },
          "restricoes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "renajud": {
            "nullable": true,
            "type": "boolean"
          },
          "rouboFurto": {
            "nullable": true,
            "type": "boolean"
          },
          "leilao": {
            "nullable": true,
            "type": "boolean"
          },
          "alienacaoFiduciaria": {
            "nullable": true,
            "type": "boolean"
          },
          "anoLicenciamento": {
            "nullable": true,
            "type": "number"
          },
          "usuarioNome": {
            "nullable": true,
            "type": "string"
          },
          "verificadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "fornecedor",
          "baseLegal",
          "justificativa",
          "situacao",
          "restricoes",
          "renajud",
          "rouboFurto",
          "leilao",
          "alienacaoFiduciaria",
          "anoLicenciamento",
          "usuarioNome",
          "verificadoEm"
        ],
        "additionalProperties": false
      },
      "Avistamento": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "observadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "registradoEm": {
            "type": "string",
            "format": "date-time"
          },
          "latitude": {
            "nullable": true,
            "type": "number"
          },
          "longitude": {
            "nullable": true,
            "type": "number"
          },
          "descricao": {
            "type": "string"
          },
          "fonte": {
            "type": "string",
            "enum": [
              "equipe_campo",
              "credor",
              "devedor",
              "outro",
              "radar"
            ]
          },
          "usuarioNome": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "observadoEm",
          "registradoEm",
          "latitude",
          "longitude",
          "descricao",
          "fonte",
          "usuarioNome"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/fipe/{tipo}/marcas": {
      "get": {
        "tags": [
          "fipe"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "carros",
                "motos",
                "caminhoes"
              ]
            },
            "in": "path",
            "name": "tipo",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "codigo": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "codigo",
                      "nome"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/fipe/{tipo}/marcas/{marca}/modelos": {
      "get": {
        "tags": [
          "fipe"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "carros",
                "motos",
                "caminhoes"
              ]
            },
            "in": "path",
            "name": "tipo",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "marca",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "codigo": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "codigo",
                      "nome"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/fipe/{tipo}/marcas/{marca}/modelos/{modelo}/anos": {
      "get": {
        "tags": [
          "fipe"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "carros",
                "motos",
                "caminhoes"
              ]
            },
            "in": "path",
            "name": "tipo",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "marca",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "modelo",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "codigo": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "codigo",
                      "nome"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/fipe": {
      "post": {
        "tags": [
          "fipe"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipo": {
                    "type": "string",
                    "enum": [
                      "carros",
                      "motos",
                      "caminhoes"
                    ]
                  },
                  "marca": {
                    "type": "string",
                    "minLength": 1
                  },
                  "modelo": {
                    "type": "string",
                    "minLength": 1
                  },
                  "ano": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "tipo",
                  "marca",
                  "modelo",
                  "ano"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValorFipe"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/verificacoes-veiculares": {
      "get": {
        "tags": [
          "veiculo"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VerificacaoVeicular"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/avistamentos": {
      "get": {
        "tags": [
          "veiculo"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Avistamento"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "veiculo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "observadoEm": {},
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "descricao": {
                    "type": "string",
                    "minLength": 5
                  },
                  "fonte": {
                    "type": "string",
                    "enum": [
                      "equipe_campo",
                      "credor",
                      "devedor",
                      "outro"
                    ]
                  }
                },
                "required": [
                  "observadoEm",
                  "descricao",
                  "fonte"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Criado"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  }
}